fix: Comment Soft Delete 적용 및 Query Method 분리
Related: #105
