+++ src/main/java/com/example/dgbackend/domain/combinationcomment/service/CombinationCommentCommandServiceImpl.java
@@ -13,9 +13,6 @@ import com.example.dgbackend.domain.combinationcomment.dto.CombinationCommentReq
 import com.example.dgbackend.domain.combinationcomment.dto.CombinationCommentResponse;
 import com.example.dgbackend.domain.combinationcomment.repository.CombinationCommentRepository;
 import com.example.dgbackend.domain.member.Member;
-import com.example.dgbackend.domain.member.repository.MemberRepository;
-import com.example.dgbackend.global.common.response.code.status.ErrorStatus;
-import com.example.dgbackend.global.exception.ApiException;
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
@@ -27,8 +24,8 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class CombinationCommentCommandServiceImpl implements CombinationCommentCommandService {
 
-    private final MemberRepository memberRepository;
     private final CombinationCommentRepository combinationCommentRepository;
+    private final CombinationCommentQueryService combinationCommentQueryService;
     private final CombinationQueryService combinationQueryService;
 
 
@@ -55,34 +52,17 @@ public class CombinationCommentCommandServiceImpl implements CombinationCommentC
         CombinationComment newComment = Optional.ofNullable(request.getParentId())
             .filter(parentId -> parentId != 0)
             .map(parentId -> toCombinationComment(combination, member, content,
-                getParentComment(request.getParentId())))
+                combinationCommentQueryService.getParentComment(request.getParentId())))
             .orElse(toCombinationComment(combination, member, content, null));
 
         return combinationCommentRepository.save(newComment);
     }
 
     @Override
-    public CombinationComment getParentComment(Long parentId) {
-        CombinationComment parentComment = getComment(parentId);
-
-        // 부모의 부모 댓글이 존재할 경우 => 에러 핸들러 실행 (대댓글까지 가능)
-        if (parentComment.getParentComment() != null) {
-            throw new ApiException(ErrorStatus._OVER_DEPTH_COMBINATION_COMMENT);
-        }
-        return parentComment;
-    }
-
-    @Override
-    public CombinationComment getComment(Long commentId) {
-        return combinationCommentRepository.findById(commentId).orElseThrow(
-            () -> new ApiException(ErrorStatus._COMBINATION_COMMENT_NOT_FOUND)
-        );
-    }
-
-    @Override
     public CombinationCommentResponse.CommentProcResult deleteComment(Long commentId) {
 
-        CombinationComment combinationComment = getComment(commentId);
+        CombinationComment combinationComment = combinationCommentQueryService.getComment(
+            commentId);
         combinationComment.deleteComment();
 
         Optional.ofNullable(combinationComment.getChildComments())
@@ -94,7 +74,8 @@ public class CombinationCommentCommandServiceImpl implements CombinationCommentC
     public CombinationCommentResponse.CommentProcResult updateComment(Long commentId,
         CombinationCommentRequest.UpdateComment request) {
 
-        CombinationComment combinationComment = getComment(commentId);
+        CombinationComment combinationComment = combinationCommentQueryService.getComment(
+            commentId);
 
         combinationComment.updateComment(request.getContent());
